+++ nengo/builder/neurons.py
@@ -1,9 +1,12 @@
+from collections import OrderedDict
+
 import numpy as np
 
 from nengo.builder import Builder, Operator, Signal
 from nengo.dists import Distribution
 from nengo.exceptions import BuildError
 from nengo.neurons import NeuronType
+from nengo.rc import rc
 from nengo.utils.numpy import is_array_like
 
 
@@ -50,7 +53,9 @@ class SimNeurons(Operator):
         super().__init__(tag=tag)
         self.neurons = neurons
 
-        self.sets = [output] + ([] if states is None else states)
+        self.states = OrderedDict() if states is None else states
+
+        self.sets = [output] + list(self.states.values())
         self.incs = []
         self.reads = [J]
         self.updates = []
@@ -63,22 +68,58 @@ class SimNeurons(Operator):
     def output(self):
         return self.sets[0]
 
-    @property
-    def states(self):
-        return self.sets[1:]
-
     def _descstr(self):
         return "%s, %s, %s" % (self.neurons, self.J, self.output)
 
     def make_step(self, signals, dt, rng):
         J = signals[self.J]
         output = signals[self.output]
-        states = [signals[state] for state in self.states]
+        states = {name: signals[sig] for name, sig in self.states.items()}
 
         def step_simneurons():
-            self.neurons.step_math(dt, J, output, *states)
+            self.neurons.step_math(dt, J, output, **states)
 
         return step_simneurons
+
+
+def get_neuron_states(model, neurontype, neurons, dtype=None):
+    """Get the initial neuron state values for the neuron type."""
+    dtype = rc.float_dtype if dtype is None else dtype
+    ens = neurons.ensemble
+    rng = np.random.RandomState(model.seeds[ens] + 1)
+    n_neurons = neurons.size_in
+
+    if isinstance(ens.initial_phase, Distribution):
+        phases = ens.initial_phase.sample(n_neurons, rng=rng)
+    else:
+        phases = ens.initial_phase
+
+    if phases.ndim != 1 or phases.size != n_neurons:
+        raise BuildError("`initial_phase` array must be 1-D of length `n_neurons`")
+
+    return neurontype.make_neuron_state(phases, model.dt, dtype=dtype)
+
+
+def sample_state_init(state_init, n_neurons, dtype=None):
+    """Sample a state init value, ensuring the correct size. """
+    dtype = rc.float_dtype if dtype is None else dtype
+
+    if isinstance(state_init, Distribution):
+        raise NotImplementedError()
+    elif is_array_like(state_init):
+        state_init = np.asarray(state_init, dtype=dtype)
+        if state_init.ndim == 0:
+            state_init = state_init * np.ones(n_neurons, dtype=dtype)
+        elif (
+            state_init.ndim == 1 and state_init.size != n_neurons or state_init.ndim > 1
+        ):
+            raise BuildError(
+                "State init array must be 0-D, or 1-D of length `n_neurons`"
+            )
+    else:
+        raise BuildError("State init must be a distribution or array-like")
+
+    return state_init
 
 
 @Builder.register(NeuronType)
@@ -104,32 +145,19 @@ def build_neurons(model, neurontype, neurons):
     Does not modify ``model.params[]`` and can therefore be called
     more than once with the same `.NeuronType` instance.
     """
-    dtype = model.sig[neurons]["in"].dtype
     n_neurons = neurons.size_in
-    state_init = neurontype.make_neuron_state(n_neurons, model.dt, dtype=dtype)
+    state_init = get_neuron_states(model, neurontype, neurons)
 
-    states = []
+    states = OrderedDict()
     for key, init in state_init.items():
         if key in model.sig[neurons]:
             raise BuildError("State name %r overlaps with existing signal name" % key)
 
-        if isinstance(init, Distribution):
-            raise NotImplementedError()
-        elif is_array_like(init):
-            init = np.asarray(init, dtype=dtype)
-            if init.ndim == 0:
-                init = init * np.ones(n_neurons, dtype=dtype)
-            elif init.ndim == 1 and init.size != n_neurons or init.ndim > 1:
-                raise BuildError(
-                    "State init array must be 0-D, or 1-D of length `n_neurons`"
-                )
-        else:
-            raise BuildError("State init must be a distribution or array-like")
-
-        model.sig[neurons][key] = Signal(
-            initial_value=init, shape=(n_neurons,), name="%s.%s" % (neurons, key)
+        states[key] = model.sig[neurons][key] = Signal(
+            initial_value=sample_state_init(init, n_neurons),
+            shape=(n_neurons,),
+            name="%s.%s" % (neurons, key),
         )
-        states.append(model.sig[neurons][key])
 
     model.add_op(
         SimNeurons(